Fixed connection edge cases
There were numerous bugs in establishing DB connections, most of them
arose from poor option handling in the pg gem. A cleaned up params hash
will get passed.

Also, many extra "require"s were removed for stability. They will be restored
in a future commit. Removing them was a scramble-for-a-fix fix for intermittent
connection issues.

Lastly, there is a bug when you run `mystic console #{env}` in the ./example/
directory. It doesn't exist if the directory is copied elsewhere and the same
command is executed. It probably occurs because the gemspec and/or gemfile is
picked up seeing as it's up a couple levels in the file structure.