Added significant stablility
The postgres extension now works! It was rewritten in a combination
of C (for the ruby wrapper) and C++ (for the Postgres and PostgresResult
wrapper). Now memory management is handled by RAII so the extension is
quite stable. A test was written to that the core database access code works.

The configuration code has been changed significantly. No longer is a connection
string generated each time a connection is to be made! It was slow and LibPQ would
just have to undo my work anyway to be able to use the connection parameters. This
way, all connection code is written in C++ and cannot be changed during runtime
(inadvertently making the library more secure).

Finally, I removed some files that were no longer needed (things like postgres.rb -
see above).