Fix symbolize bug & write Postgres-based queue
Fixed a bug in Object#symbolize where it was possible
that a non-string could be sent :to_sym & disabled the
symbolizing of strings that are values in Hashes.

Wrote a Postgres queue based on QueueClassic that does
essentially the same thing, except it is integrated with
Mystic so that I don't have to install the pg gem.

What's left is writing the actual socket code so that Mystic
can implement a wait-on-notify behavior.